Refactor classes in Flight ODBC into type_fwd.h
