+++ cpp/src/arrow/flight/sql/odbc/odbc_impl/spi/result_set.h
@@ -23,12 +23,11 @@
 #include "arrow/flight/sql/odbc/odbc_impl/platform.h"
 
 #include "arrow/flight/sql/odbc/odbc_impl/types.h"
+#include "arrow/flight/sql/odbc/odbc_impl/type_fwd.h"
 
 #include <sqltypes.h>
 
 namespace arrow::flight::sql::odbc {
-
-class ResultSetMetadata;
 
 class ResultSet {
  protected:
