+++ cpp/src/arrow/flight/sql/odbc/odbc_impl/spi/statement.h
@@ -22,11 +22,9 @@
 #include <optional>
 #include <vector>
 
+#include "arrow/flight/sql/odbc/odbc_impl/type_fwd.h"
+
 namespace arrow::flight::sql::odbc {
-
-class ResultSet;
-
-class ResultSetMetadata;
 
 /// \brief High-level representation of an ODBC statement.
 class Statement {
